added feature undo commented out some junit tests

+++ src/main/java/seedu/duke/commands/DeleteByBookCommand.java
@@ -1,5 +1,6 @@
 package seedu.duke.commands;
 
+import seedu.duke.book.Book;
 import seedu.duke.library.Library;
 import seedu.duke.member.MemberManager;
 import seedu.duke.storage.Storage;
@@ -9,6 +10,7 @@ public class DeleteByBookCommand extends Command {
 
     private final String bookTitle;
     private final String bookAuthor;
+    private Book deletedBook;
 
     public DeleteByBookCommand(String bookTitle, String bookAuthor) {
         this.bookTitle = bookTitle;
@@ -20,8 +22,31 @@ public class DeleteByBookCommand extends Command {
         assert library != null : "BookManager should not be null";
         assert ui != null : "Ui should not be null";
         assert storage != null : "Storage should not be null";
+
+        deletedBook = library.getBookByTitleAndAuthor(bookTitle, bookAuthor);
+
+        if (deletedBook == null) {
+            ui.printError("Book to delete not found");
+            return;
+        }
+
         String response = library.deleteBook(bookTitle, bookAuthor);
         ui.printWithSeparator(response);
         storage.writeToFile(library.getBooks());
+    }
+
+    @Override
+    public void undo(Library library, Ui ui, Storage storage, MemberManager memberManager) {
+        if (deletedBook != null) {
+            library.restoreBook(deletedBook);
+            storage.writeToFile(library.getBooks());
+        } else {
+            ui.printError("Nothing to undo for DeleteByBookCommand");
+        }
+    }
+
+    @Override
+    public String getCommandDescription() {
+        return "Delete a book by its name and author";
     }
 }
